requirements.txt
numpy~=1.24.2
pandas~=1.5.3
enum34~=1.1.10
opencv-python~=4.7.0.72
pyrealsense2~=2.53.1.4623
torch~=1.13.1
Pillow~=9.0.1
yolov5~=7.0.9
PyYAML~=6.0
torchvision~=0.14.1
matplotlib~=3.7.1
seaborn~=0.12.2
requests~=2.25.1
tqdm~=4.65.0
psutil~=5.9.0
thop~=0.1.1.post2209072238